fix: Create Flask request context in CLI send_any_notifications command
Copied from ckanext-harvest's CLI commands that ultimately send emails.

We use ckan.lib.base.render to render email contents as HTML, and this function requires a request context. All other kinds of emails we send are prompted by a request (e.g. clicking on a subscribe or management link) so the request context exists, but when we prompt email sending with a CLI command, we need to create this context beforehand.

It looks a bit odd to use the 'test_request_context' method here, but Flask prefers that 'request_context' is only used when an actual request is handled and external code uses 'test_request_app'.